Use AUDIO_BECOMING_NOISY event; remove headset only option
The AUDIO_BECOMING_NOISY intent seems to be broadcast before the
HEADSET_PLUG intent. In some cases, the HEADSET_PLUG intent fires so
late that audio starts playing through the speaker..
AUDIO_BECOMING_NOISY also handles Bluetooth.

The headset only feature was useful to prevent accidental "noisy" playback,
however it is cumbersome to implement with the Android audio API being the mess
that it is. Thus I've removed it.