Remove programming-created media button receiver
This was an optimization (since manifest-defined receivers have to be
constructed again for each broadcast) that doesn't seem to help all that much.
I mainly removed it because it makes using the Froyo registerMediaButton API
harder (impossible?) to use.

Also change receiver priority to 999, since apparently >=1000 should only be
used by the system.