Prevent redundant cover work in CoverView
There were two main cases where we would generate covers: when querySongs is
called and when onSizeChanged is called. querySongs was sometimes called before
the initial layout had occured and the CoverView's size determined (usually
when the service had already been running when a new PlaybackActivity is
created). onSizeChanged was sometimes called before the service had been
initialized. Both of these cases are not useful, and in the former case,
harmful, since the default cover art could be shown when the view should have
been blank. So this removes generation in onSizeChanged except when querySongs
had been called before the layout occurs, which elminates the redundant cases.