Fix a crash bug due to mCurrentSong == null This happens if: * Notification is set to always * The user empties the queue
This causes activeSongReplaced() to be called, so setCurrentSong(0); will be executed while mCurrentSong is already null.
Updating the notification in this state crashes with a null pointer.

However: The 'dangling' notification will have wrong information in this state - but it is still better than displaying an empty one as it just looks butt ugly.
We whould probably, at some point, design an 'empty notification view'