DragSortController: Check for null MotionEvent arguments (#202)
This fixes a rare crash on Samsung Galaxy S6 devices. The null arguments
appear to be caused by the GestureDetector not receiving an ACTION_DOWN
event before receiving an ACTION_MOVE event.

As we are unable to reproduce it, this commit work around the issue. It
might cause some scrolling issues in the cases that cause the issue, but
better than crashing.