Fix out-of-date settings shown after import
We use the PreferenceFragment framework provided by Android to
display settings and allow the user to modify them. The API doesn't
support dynamic preference changes from a non-user source, such as
loading the preferences from a file.

Previously, after importing the settings from a file the user would go
back to the PreferenceActivity and would see the old (pre-import)
values, even though the saved settings had been correctly loaded.

This commit provides a stop-gap solution to the issue by
finishing the PreferenceActivity when the user imports settings, the
rationale being that forcing the user to re-navigate to the settings,
while slightly annoying, is much better than making them think we hosed
their data.