Add Gradle wrapper v2.2.1
A gradle wrapper script allows a developer to download and use the
correct version of gradle for Vanilla without needing to install it on
their machine. Vanilla uses v1.3.0 of the Android Gradle plugin, which
requires Gradle v2.2.1.

This commit adds the gradle wrapper, generated using

```
gradle wrapper --gradle-version 2.2.1
```

This command generates a gradle-wrapper.properties that links to the
binary distribution of Gradle. For better integration with IDE's, this
commit changes the URL to '*-all.zip'

More information about the Gradle wrapper can be found in the
[Gradle User Guide - Wrapper](https://docs.gradle.org/current/userguide/gradle_wrapper.html)