Introduce FastScrollGuardListView
This is a hack to work around androids stupid 48dip fast-scroll area: The framework will hijack any scroll events which happen
near (=48 dp) the fast scroll bar. We do not want this as it clashes with our expand buttons.

This class intercepts events in the problematic area and rewrites them to a fake event to pretend that the touch did not happen in the fastscroll area