Add media/audio to library default blacklist
The media/audio folder can be used as an alternative combined folder to add custom sounds to various android choosers.

This folder may hold for example the following subdirectories:
* alarms (/sdcard/Alarms alternative)
* notifications (/sdcard/Notifications alternative)
* ringtones (/sdcard/Ringtones alternative)
* ui (dont know if /sdcard/Ui is also such a directory)

All of these are scanned by Android in the same way like the other special directories.
As they hold e.g. click or short alarm sounds or like that they are not needed in music player ;).