Custom drawn icons
These icons bring the Sony Walkman aesthetics to Vanilla Music,
however these are made from 100% original vector assets that are also a
part of this update

Disclaimer: it has been reported that the icons might look inconsistent
on some Android flavours. I have only tested this on stock versions of
Android 11 and 12.